Move the drawing commands to GPUCommonHW
